move nlx code to separate component
