+++ src/openzaak/js/components/admin/external-services/external-form.js
@@ -1,17 +1,9 @@
-import React, {useContext, useState} from "react";
-import {TextInput, CheckBoxInputLabel} from "../../../forms/inputs";
+import React from "react";
+import {TextInput} from "../../../forms/inputs";
 import {API_TYPES} from "../../../forms/constants";
 import {SelectInput} from "./select";
-import {ConstantsContext} from "./context";
 import {AuthType} from "./auth-type";
-
-function getInitialOrganization (nlx_url, outway) {
-    if (!outway || !nlx_url || !nlx_url.startsWith(outway)) {
-        return '';
-    }
-    const path = nlx_url.slice(outway.length);
-    return path.split('/')[0];
-}
+import {Nlx} from "./nlx";
 
 
 function ExternalForm(props) {
@@ -20,20 +12,6 @@ function ExternalForm(props) {
 
     const id_prefix = (field) => `id_form-${index}-${field}`;
     const name_prefix = (field) => `form-${index}-${field}`;
-
-    // nlx calculation
-    const { nlxOutway, nlxChoices } = useContext(ConstantsContext);
-    const [ isNlx, toggleNlx ] = useState(Boolean(values.nlx));
-    const initialOrganization = getInitialOrganization(values.nlx, nlxOutway);
-    const [ selectedOrganization, setSelectedOrganization ] = useState(initialOrganization);
-    const [ selectedService, setSelectedService ] = useState(values.nlx);
-    const organizationChoices = Object.keys(nlxChoices).map((value) => [value, value]);
-    const services = nlxChoices[selectedOrganization];
-    const serviceChoices = selectedOrganization
-        ? Object.keys(services).map((service) => [service, services[service].service_name])
-        : [];
-    const isNlxDisabled = !nlxOutway;
-
     const isEven = (index % 2) === 0;
 
     return (
@@ -77,52 +55,7 @@ function ExternalForm(props) {
 
             {/*nlx*/}
             <td className='external-form__field'>
-                <div title={isNlxDisabled ? "NLW outway must be set up" : null}>
-                <CheckBoxInputLabel
-                    name={name_prefix('is_nlx')}
-                    value={'is_nlx'}
-                    label={'Use NLX?'}
-                    id={id_prefix('is_nlx')}
-                    checked={isNlx}
-                    onChange={() => toggleNlx(!isNlx)}
-                    disabled={isNlxDisabled}
-                />
-                </div>
-
-                {(isNlx && !isNlxDisabled) ? (
-                    <>
-                        <div className='external-form__group'>
-                            <label
-                                htmlFor={id_prefix('nlx_organizations')}
-                                className='external-form__label'
-                            >Organization:</label>
-                            <SelectInput
-                                choices={organizationChoices}
-                                name={name_prefix('nlx_organizations')}
-                                id={id_prefix('nlx_organizations')}
-                                initialValue={selectedOrganization}
-                                onChange={(organization) => setSelectedOrganization(organization)}
-                                classes="external-form__field--wide"
-                            />
-                        </div>
-
-                        <div className='external-form__group'>
-                            <label
-                                htmlFor={id_prefix('nlx')}
-                                className='external-form__label'
-                            >Service:</label>
-                            <SelectInput
-                                choices={serviceChoices}
-                                name={name_prefix('nlx')}
-                                id={id_prefix('nlx')}
-                                initialValue={selectedService}
-                                onChange={(service) => setSelectedService(service)}
-                                classes="external-form__field--wide"
-                            />
-                        </div>
-                    </>
-                    ): null
-                }
+                <Nlx index={index} data={data} />
             </td>
 
             {/*OAS*/}
